+++ noodles-vcf/src/record/builder.rs
@@ -276,19 +276,16 @@ impl Builder {
     /// use noodles_vcf::{
     ///     self as vcf,
     ///     header::format::Key,
-    ///     record::{
-    ///         genotypes::{genotype::{field::Value, Field}, Genotype},
-    ///         Genotypes, Position,
-    ///     },
+    ///     record::{genotypes::genotype::field::Value, Genotypes, Position},
     /// };
     ///
     /// let keys = "GT:GQ".parse()?;
     /// let genotypes = Genotypes::new(
     ///     keys,
-    ///     vec![Genotype::try_from(vec![
-    ///         Field::new(Key::Genotype, Some(Value::String(String::from("0|0")))),
-    ///         Field::new(Key::ConditionalGenotypeQuality, Some(Value::Integer(13))),
-    ///     ])?],
+    ///     vec![[
+    ///         (Key::Genotype, Some(Value::String(String::from("0|0")))),
+    ///         (Key::ConditionalGenotypeQuality, Some(Value::Integer(13))),
+    ///     ].into_iter().collect()],
     /// );
     ///
     /// let record = vcf::Record::builder()