+++ noodles-vcf/src/writer.rs
@@ -184,10 +184,7 @@ mod tests {
         use crate::{
             header::format::Key,
             record::{
-                genotypes::{
-                    genotype::{field::Value, Field},
-                    Genotype, Keys,
-                },
+                genotypes::{genotype::field::Value, Keys},
                 Genotypes,
             },
         };
@@ -196,10 +193,12 @@ mod tests {
 
         let genotypes = Genotypes::new(
             Keys::try_from(vec![Key::Genotype, Key::ConditionalGenotypeQuality])?,
-            vec![Genotype::try_from(vec![
-                Field::new(Key::Genotype, Some(Value::String(String::from("0|0")))),
-                Field::new(Key::ConditionalGenotypeQuality, Some(Value::Integer(13))),
-            ])?],
+            vec![[
+                (Key::Genotype, Some(Value::String(String::from("0|0")))),
+                (Key::ConditionalGenotypeQuality, Some(Value::Integer(13))),
+            ]
+            .into_iter()
+            .collect()],
         );
 
         let record = Record::builder()